requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
numpy>=1.21.2
matplotlib>=3.4.3
tqdm>=4.62.2
scikit-learn>=0.24.2 